[plugin] NewsDownloader: fix missing elements in the 'Add feeds' window  (#13415)
The current code will replace "empty" UI elements with a line.
However it is relying on the truthiness of the value of the field.
If the value is False, for instance for the "download image" item,
then the UI element for that field will be replaced with a line in the
UI.